requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
Babel==2.9.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
environs==8.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==3.26.0
multidict==6.1.0
nodeenv==1.9.1
packaging==24.2
prisma==0.15.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
tomlkit==0.13.2
typing_extensions==4.12.2
yarl==1.18.3
